requirements.txt
flask==3.0.2
werkzeug==3.0.1
pdfplumber==0.11.4
docx2txt==0.8
bert-extractive-summarizer==0.10.1
transformers>=4.35.0

torch
pke==2.0.0
spacy>=3.7.0
nltk>=3.8.1
